Someone clear idea of how to evaluate the ast and perform metaprogramming

+++ src/types.rs
@@ -1,2 +1,2 @@
 pub use i32 as Numeric;
-pub use i32 as Id;
+pub use usize as Id;
